# src/Middagsasen.Planner.Api/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Middagsasen.Planner.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Competencies": {
      "get": {
        "tags": [
          "Competencies"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetencyResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetencyResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetencyResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competencies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/{id}": {
      "get": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/user/{userId}": {
      "get": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserCompetencyResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserCompetencyResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserCompetencyResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/user": {
      "post": {
        "tags": [
          "Competencies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCompetencyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCompetencyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserCompetencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/user/{userCompetencyId}/approve": {
      "put": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "userCompetencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveCompetencyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveCompetencyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveCompetencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/user/{userCompetencyId}": {
      "delete": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "userCompetencyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCompetencyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/{id}/approvers/{userId}": {
      "post": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyApproverResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyApproverResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyApproverResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Competencies/approvers/{approverId}": {
      "delete": {
        "tags": [
          "Competencies"
        ],
        "parameters": [
          {
            "name": "approverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApproveCompetencyRequest": {
        "type": "object",
        "properties": {
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "CompetencyApproverResponse": {
        "required": [
          "id",
          "userId",
          "fullName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          }
        }
      },
      "CompetencyRequest": {
        "required": [
          "name",
          "hasExpiry"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasExpiry": {
            "type": "boolean"
          },
          "resourceTypeIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "resourceTypeCompetencies": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ResourceTypeCompetencyRequest"
            }
          }
        }
      },
      "CompetencyResourceTypeResponse": {
        "required": [
          "resourceTypeId",
          "resourceTypeName",
          "minimumRequired"
        ],
        "type": "object",
        "properties": {
          "resourceTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "resourceTypeName": {
            "type": "string"
          },
          "minimumRequired": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CompetencyResponse": {
        "required": [
          "id",
          "name",
          "hasExpiry",
          "inactive",
          "resourceTypes",
          "approvers"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasExpiry": {
            "type": "boolean"
          },
          "inactive": {
            "type": "boolean"
          },
          "resourceTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompetencyResourceTypeResponse"
            }
          },
          "approvers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompetencyApproverResponse"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ResourceTypeCompetencyRequest": {
        "required": [
          "resourceTypeId",
          "minimumRequired"
        ],
        "type": "object",
        "properties": {
          "resourceTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "minimumRequired": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UserCompetencyRequest": {
        "required": [
          "userId",
          "competencyId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "competencyId": {
            "type": "integer",
            "format": "int32"
          },
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "UserCompetencyResponse": {
        "required": [
          "id",
          "userId",
          "userFullName",
          "competencyId",
          "competencyName",
          "approved",
          "isExpired",
          "created"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "userFullName": {
            "type": "string"
          },
          "competencyId": {
            "type": "integer",
            "format": "int32"
          },
          "competencyName": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          },
          "approvedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "approvedById": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "approvedByName": {
            "type": [
              "null",
              "string"
            ]
          },
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isExpired": {
            "type": "boolean"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT-token fra /api/authentication/authenticate. Lim inn kun selve tokenet (uten \"Bearer \").",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}